hp2ps/Utilities.c: add extern parameter types for malloc and realloc for C23
use portable C types!

+++ utils/hp2ps/Utilities.c
@@ -3,7 +3,7 @@
 #include <string.h>
 #include "Error.h"
 
-extern void* malloc(long unsigned int);
+extern void* malloc(size_t);
 
 char*
 Basename(char *name)
@@ -89,7 +89,7 @@ void *
 xrealloc(void *p, size_t n)
 {
     void *r;
-    extern void *realloc(void *, long unsigned int);
+    extern void *realloc(void *, size_t);
 
     r = realloc(p, n);
     if (!r) {
